extract package for database settings in jbb-system-web
