+++ domain-web/jbb-system-web/src/main/java/org/jbb/system/web/database/data/FormDatabaseSettings.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016 the original author or authors.
+ * Copyright (C) 2017 the original author or authors.
  *
  * This file is part of jBB Application Project.
  *
@@ -8,10 +8,10 @@
  *        http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package org.jbb.system.web.base.data;
+package org.jbb.system.web.database.data;
 
 import org.jbb.system.api.database.DatabaseSettings;
-import org.jbb.system.web.base.form.DatabaseSettingsForm;
+import org.jbb.system.web.database.form.DatabaseSettingsForm;
 
 import lombok.experimental.Delegate;
