+++ domain-web/jbb-system-web/src/test/java/org/jbb/system/web/database/data/FormDatabaseSettingsTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016 the original author or authors.
+ * Copyright (C) 2017 the original author or authors.
  *
  * This file is part of jBB Application Project.
  *
@@ -8,9 +8,9 @@
  *        http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package org.jbb.system.web.base.data;
+package org.jbb.system.web.database.data;
 
-import org.jbb.system.web.base.form.DatabaseSettingsForm;
+import org.jbb.system.web.database.form.DatabaseSettingsForm;
 import org.junit.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
